[tests][m]: tests for each role
